clk: Guard orphan_list under DEBUG_FS
The orphan_list is a structure pointer which holds the records of
the subtrees with no root list and dumps them in a debug directory
while populating the directory on each early boot.

Since this spits out an "-unused_variable" warning if DEBUG_FS is
disabled, we should be guarding this with an ifdef.

Fixes f114a36246812b5c06b0a6066412215e45b3ac8c
(clk: Evict unregistered clks from parent caches)

Change-Id: I67f5864e2daa2de90d443d717d72c0b9a32465be
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>